requirements.txt
fastapi==0.115.6
pandas==2.2.3
python-dotenv==1.0.1
torch==2.4.1
transformers==4.44.1
uvicorn==0.30.3